updated the structure for Yntraa
