+++ testcases/automated_cases/compute_suite/test_snapshots.py
